refactor: 회원가입 텍스트 필드 코드 개선
